backend/requirements.txt
# =====================================
# CRYPTO VIZ - Backend Requirements (Minimal)
# =====================================

# Bootstrap Stack (requis EPITECH)
pandas==2.1.3
duckdb==0.9.2

# FastAPI Core
fastapi==0.104.1
uvicorn==0.24.0

# Basic HTTP
requests==2.31.0

# Kafka
kafka-python==2.0.2

# Database drivers
psycopg2-binary==2.9.9

# Data processing
numpy==1.24.3
python-dateutil==2.8.2

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8

# Security
cryptography==41.0.7

# Cache
redis==5.0.1
hiredis==2.2.3

# WebSocket
websockets==12.0